requirements.txt
absl-py==0.9.0
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
astor==0.8.1
beautifulsoup4==4.8.0
Brlapi==0.6.6
certifi==2019.6.16
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.0.3
django-crispy-forms==1.8.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Flask==1.1.1
gast==0.3.3
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.9.2
idna==2.8
indicator-cpufreq==0.2.2
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.1
netifaces==0.10.4
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
pandas==1.0.0
pexpect==4.2.1
Pillow==5.1.0
postgres==3.0.0
protobuf==3.11.3
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
pyRFC3339==1.0
pyTelegramBotAPI==3.6.6
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
soupsieve==1.9.3
sqlparse==0.3.1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tb-nightly==1.14.0a20190301
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.3
virtualenv==16.7.9
wadllib==1.3.2
Werkzeug==1.0.0
wrapt==1.12.1
xkit==0.0.0
zope.interface==4.3.2
